# Conflicts:
#	src/functionalTest/resources/features/opalMode/defendantAccounts/DefendantAccountSearchFeatureToggles.feature
#	src/functionalTest/resources/features/opalMode/featureToggles/FinesServiceRelease1cFeatureToggles.feature
#	src/main/java/uk/gov/hmcts/opal/util/FeatureFlags.java
#	src/main/resources/application.yaml
#	src/test/java/uk/gov/hmcts/opal/disco/opal/JsonSchemaValidationServiceTest.java

+++ src/test/java/uk/gov/hmcts/opal/service/DefendantAccountSearchRequestValidatorTest.java
@@ -107,7 +107,7 @@ class DefendantAccountSearchRequestValidatorTest {
     private PostDefendantAccountSearchRequestDefendantAccount request(boolean consolidationSearch) {
         return PostDefendantAccountSearchRequestDefendantAccount.builder()
             .activeAccountsOnly(true)
-            .businessUnitIds(List.of(78))
+            .businessUnitIds(List.of((short) 78))
             .consolidationSearch(consolidationSearch)
             .referenceNumber(new DefendantAccountSearchReferenceNumberDefendantAccount()
                 .organisation(false)
